+++ sickrage/providers/torrent/torrentday.py
@@ -22,8 +22,8 @@ from __future__ import unicode_literals
 
 import re
 
-import requests
 import sickrage
+from requests.utils import dict_from_cookiejar
 from sickrage.core.caches.tv_cache import TVCache
 from sickrage.providers import TorrentProvider
 
@@ -33,21 +33,19 @@ class TorrentDayProvider(TorrentProvider):
 
         super(TorrentDayProvider, self).__init__("TorrentDay", 'classic.torrentday.com', True)
 
-        self.supports_backlog = True
-
-        self._uid = None
-        self._hash = None
         self.username = None
         self.password = None
-        self.ratio = None
+
         self.freeleech = False
         self.minseed = None
         self.minleech = None
 
+        self.enable_cookies = True
+
         self.cache = TVCache(self, min_time=10)
 
         self.urls.update({
-            'login': '{base_url}/torrents/'.format(base_url=self.urls['base_url']),
+            'login': '{base_url}/t'.format(base_url=self.urls['base_url']),
             'search': '{base_url}/V3/API/API.php'.format(base_url=self.urls['base_url']),
             'download': '{base_url}/download.php/%s/%s'.format(base_url=self.urls['base_url'])
         })
@@ -58,39 +56,29 @@ class TorrentDayProvider(TorrentProvider):
                            'RSS': {'c2': 1, 'c26': 1, 'c7': 1, 'c24': 1, 'c14': 1}}
 
     def login(self):
-
-        if any(requests.utils.dict_from_cookiejar(sickrage.srCore.srWebSession.cookies).values()):
+        cookie_dict = dict_from_cookiejar(self.cookie_jar)
+        if cookie_dict.get('uid') and cookie_dict.get('pass'):
             return True
 
-        if self._uid and self._hash:
-            requests.utils.add_dict_to_cookiejar(sickrage.srCore.srWebSession.cookies, self.cookies)
-        else:
-
-            login_params = {'username': self.username,
-                            'password': self.password,
-                            'submit.x': 0,
-                            'submit.y': 0}
-
-            try:
-                response = sickrage.srCore.srWebSession.post(self.urls['login'], data=login_params, timeout=30).text
-            except Exception:
-                sickrage.srCore.srLogger.warning("[{}]: Unable to connect to provider".format(self.name))
-                return False
-
-            if re.search('You tried too often', response):
-                sickrage.srCore.srLogger.warning("Too many login access attempts")
-                return False
-
-            try:
-                    self._uid = requests.utils.dict_from_cookiejar(sickrage.srCore.srWebSession.cookies)['uid']
-                    self._hash = requests.utils.dict_from_cookiejar(sickrage.srCore.srWebSession.cookies)['pass']
-                    self.cookies = {'uid': self._uid, 'pass': self._hash}
-                    return True
-            except Exception:
-                pass
-
-            sickrage.srCore.srLogger.warning("Unable to obtain cookie")
+        if not self.add_cookies_from_ui():
             return False
+
+        login_params = {'username': self.username, 'password': self.password, 'submit.x': 0, 'submit.y': 0}
+
+        response = sickrage.srCore.srWebSession.post(self.urls['login'], data=login_params)
+        if response.status_code != 200:
+            sickrage.srCore.srLogger.warning("[{}]: Unable to connect to provider".format(self.name))
+            return False
+
+        if re.search('You tried too often', response.text):
+            sickrage.srCore.srLogger.warning("Too many login access attempts")
+            return False
+
+        if not dict_from_cookiejar(sickrage.srCore.srWebSession.cookies).get('uid') in response.text:
+            sickrage.srCore.srLogger.warning("Failed to login, check your cookies")
+            return False
+
+        return True
 
     def search(self, search_params, search_mode='eponly', epcount=0, age=0, epObj=None):